Move TextGenerationNode type and helper to actions/index.ts
