+++ packages/data-type/src/node/actions/text-generation.ts
@@ -1,6 +1,5 @@
 import { LanguageModel } from "@giselle-sdk/language-model";
 import { z } from "zod";
-import { NodeBase } from "../base";
 
 export const TextGenerationContent = z.object({
 	type: z.literal("textGeneration"),
@@ -9,18 +8,6 @@ export const TextGenerationContent = z.object({
 });
 export type TextGenerationContent = z.infer<typeof TextGenerationContent>;
 
-export const TextGenerationNode = NodeBase.extend({
-	type: z.literal("action"),
-	content: TextGenerationContent,
-});
-type TextGenerationNode = z.infer<typeof TextGenerationNode>;
-
-export function isTextGenerationNode(
-	args?: unknown,
-): args is TextGenerationNode {
-	const result = TextGenerationNode.safeParse(args);
-	return result.success;
-}
 
 export const TextGenerationContentReference = z.object({
 	type: TextGenerationContent.shape.type,
